put configs in each pkg to fix coverage
